added the filter buttons and plots to all homes
