+++ src/routes/plot/+page.svelte
@@ -1,189 +1,5 @@
 
-<!-- 
-<script>
-  // Import the Plot component
-  import Plot from '$lib/components/Plot.svelte';
-  import { onMount } from 'svelte';
-  import * as PlotLibrary from '@observablehq/plot';
 
-  // Define the plot and data variables
-  let plot;
-  let fetchedData;
-
-  // Fetch data from the API
-  async function fetchData() {
-    const token = "bmMzMjM2cGd0QHN0dWRlbnRzLm51bG9uZG9uLmFjLnVrOmM2ODU0NThhM2EyZGY2ZDI4YzRlNmQwN2FjY2U0NDRkNDViZGU1Mjk="; // Replace with your actual token
-    const response = await fetch('https://epc.opendatacommunities.org/api/v1/domestic/search?&size=100', {
-      headers: {
-        'Authorization': `Basic ${token}`,
-        'Accept': 'application/json'
-      }
-    });
-    const data = await response.json();
-    return  data["rows"]; ;
-  }
-
-  // Load data on mount
-  onMount(async () => {
-    fetchedData = await fetchData();
-    console.log(fetchedData);
-
-    const propertyTypeColumn = fetchedData.map(row => row["property-type"]);
-
-    // Preprocess the data to get counts of each property type
-    const propertyTypeCounts = {};
-    propertyTypeColumn.forEach(propertyType => {
-      propertyTypeCounts[propertyType] = (propertyTypeCounts[propertyType] || 0) + 1;
-    });
-
-    // Convert the counts object into an array of objects for Plotly
-    const plotData = Object.keys(propertyTypeCounts).map(propertyType => ({
-      'property-type': propertyType,
-      'count': propertyTypeCounts[propertyType]
-    }));
-
-    // Now, use plotData to create the plot
-    plot = PlotLibrary.plot({
-      marks: [
-        PlotLibrary.barY(plotData, {
-          x: "property-type",
-          y: "count",
-          fill: "steelblue"
-        })
-      ],
-      width: 400,
-      height: 200
-    });
-  });
-</script>
-
-{#if fetchedData}
-  Display the Plot component and pass the plot object
-  <div class="plot-container">
-    <Plot {plot} />
-  </div>
-{:else}
-  <p>Loading...</p>
-{/if}
-
-<style>
-  /* Add any custom styling for the plot container */
-  .plot-container {
-    /* max-width: 900px;
-    margin: 0 auto; */
-    width: 100%;
-    height: 100%;
-  }
-</style> -->
-<!-- <script>
-  // Import the Plot component
-  import Plot from '$lib/components/Plot.svelte';
-  import { onMount } from 'svelte';
-  import * as PlotLibrary from '@observablehq/plot';
-
-  // Define the plot and data variables
-  let plotPropertyType;
-  let plotTenure;
-  let fetchedData;
-
-  // Fetch data from the API
-  async function fetchData() {
-    const token = "bmMzMjM2cGd0QHN0dWRlbnRzLm51bG9uZG9uLmFjLnVrOmM2ODU0NThhM2EyZGY2ZDI4YzRlNmQwN2FjY2U0NDRkNDViZGU1Mjk="; // Replace with your actual token
-    const response = await fetch('https://epc.opendatacommunities.org/api/v1/domestic/search?&size=100', {
-      headers: {
-        'Authorization': `Basic ${token}`,
-        'Accept': 'application/json'
-      }
-    });
-    const data = await response.json();
-    return data["rows"];
-  }
-
-  // Load data on mount
-  onMount(async () => {
-    fetchedData = await fetchData();
-    console.log(fetchedData);
-
-    const propertyTypeColumn = fetchedData.map(row => row["property-type"]);
-    const tenureColumn = fetchedData.map(row => row["tenure"]);
-
-    // Preprocess the data to get counts of each property type
-    const propertyTypeCounts = {};
-    propertyTypeColumn.forEach(propertyType => {
-      propertyTypeCounts[propertyType] = (propertyTypeCounts[propertyType] || 0) + 1;
-    });
-
-    // Convert the counts object into an array of objects for Plotly
-    const plotPropertyTypeData = Object.keys(propertyTypeCounts).map(propertyType => ({
-      'property-type': propertyType,
-      'count': propertyTypeCounts[propertyType]
-    }));
-
-    // Preprocess the data to get counts of each tenure
-    const tenureCounts = {};
-tenureColumn.forEach(tenure => {
-  // Include only specific tenures
-  if (tenure === "Owner-occupied" || tenure === "Rented (social)" || tenure === "Rented (private)") {
-    tenureCounts[tenure] = (tenureCounts[tenure] || 0) + 1;
-  }
-});
-
-// Convert the counts object into an array of objects for Plotly
-const plotTenureData = Object.keys(tenureCounts).map(tenure => ({
-  'tenure': tenure,
-  'count': tenureCounts[tenure]
-}));
-
-    // Create the plot for property type
-    plotPropertyType = PlotLibrary.plot({
-      marks: [
-        PlotLibrary.barY(plotPropertyTypeData, {
-          x: "property-type",
-          y: "count",
-          fill: "steelblue"
-        })
-      ],
-      width: 400,
-      height: 200
-    });
-
-    // Create the plot for tenure
-    plotTenure = PlotLibrary.plot({
-      marks: [
-        PlotLibrary.barY(plotTenureData, {
-          x: "tenure",
-          y: "count",
-          fill: "steelblue"
-        })
-      ],
-      width: 400,
-      height: 200
-    });
-  });
-</script>
-
-
-{#if fetchedData}
-   Display the Plot components and pass the plot objects -->
-  <!-- <div class="plot-container">
-    <h2>Property Type</h2>
-    <Plot {plotPropertyType} />
-  </div>
-  <div class="plot-container">
-    <h2>Tenure</h2>
-    <Plot {plotTenure} />
-  </div>
-{:else}
-  <p>Loading...</p>
-{/if} --> 
-
-<!-- <style>
-  /* Add any custom styling for the plot container */
-  /* .plot-container {
-    width: 100%;
-    height: 100%;
-  } */
-</style> -->
 
 <script>
   // Import the Plot component
